alx_travel_app/requirement.txt
amqp==5.3.1
anyjson==0.3.3
apturl==0.5.2
asgiref==3.8.1
asn1crypto==1.4.0
bcrypt==3.2.0
billiard==4.2.1
blinker==1.4
Brlapi==0.8.3
celery==5.5.2
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.9
distro==1.7.0
distro-info===1.1build1
Django==5.2.1
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
drf-yasg==1.21.10
drf-yasg2==1.19.4
duplicity==0.8.21
ephem==4.1.3
evdev==1.4.0
fasteners==0.14.1
filelock==3.18.0
future==0.18.2
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.6
kazam==1.4.5
keyring==23.5.0
keyrings.alt==4.1.0
kombu==5.5.3
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libevdev==0.5
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
mysql-connector-python==9.3.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==25.0
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.3.8
prompt-toolkit==3.0.28
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0
python-dateutil==2.9.0.post0
python-debian===0.1.43ubuntu1
python-memcached==1.58
pytz==2022.1
pyudev==0.22.0
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-unixsocket==0.2.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
SecretStorage==3.3.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
ssh-import-id==5.11
systemd-python==234
typing_extensions==4.13.2
tzdata==2025.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
vine==5.1.0
virtualenv==20.31.2
wadllib==1.3.6
wcwidth==0.2.5
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0
